requirements.txt
# CR RTIS Analysis Tool - Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.121.1
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Data Processing
polars==1.35.2
pandas==2.3.3
numpy==2.3.4
pyarrow==22.0.0

# Excel Export
xlsxwriter==3.2.0
openpyxl==3.1.5

# PDF Generation
matplotlib==3.10.7
reportlab==4.4.4

# Utilities
python-dotenv==1.2.1
requests==2.32.5

# Testing
pytest==9.0.2

# Date/Time
python-dateutil==2.9.0.post0
pytz==2025.2
mysql-connector-python==8.4.0
